+++ CHANGELOG.md
@@ -5,7 +5,26 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.2.8 - Unreleased
+## Version 0.2.9 - Unreleased
+
+
+### Added
+
+* Added `vidid` argument to `CocoDataset.images` 1D object mixin.
+* Added `trackid` argument to `CocoDataset.annots` 1D object mixin.
+
+### Fixed
+
+* `perterb_coco` now correctly perterbs segmentations
+
+### Changed
+
+* Changed argument name from `gsize` to `image_size` in toydata methods.
+  Backwards compatibility is preserved via kwargs for now, but `gsize` will be
+  deprecated in the future.
+
+
+## Version 0.2.8 - Released 2021-08-02
 
 
 ### Added